refactor: factory class 제거
